feat : devide function to Cookie.jsx
